- Bug hunting
git-svn-id: svn://mirror.orbiter-radio.co.uk/D3D9client/trunk@514 c0f5cd98-1510-604a-8459-bf7538b8c9c2

+++ Orbitersdk/D3D9Client/Surfmgr2.h
@@ -57,7 +57,7 @@ protected:
 private:
 	INT16 *ElevationData () const;
 	double GetMeanElevation (const INT16 *elev) const;
-	float fixinput(float,int);
+	float fixinput(double, int);
 	D3DXVECTOR4 MicroTexRange(SurfTile *pT, int lvl) const;
 
 	D3DXVECTOR2 MicroRep[3];
